Refactoring a bit.
