+++ src/main/python/rlbot/setup_manager.py
@@ -2,16 +2,15 @@ import msvcrt
 import multiprocessing as mp
 import os
 import queue
-import time
 import sys
+import time
 
 # Make sure the flatbuffers dir can be located on sys.path so that the generated files can find it.
 # TODO: Use pip for flatbuffers if they ever get their act together: https://github.com/google/flatbuffers/issues/4507
 sys.path.insert(0, os.path.realpath(os.path.join(os.path.dirname(os.path.abspath(__file__)), "pylibs")))
 
+from rlbot.botmanager.helper_process_manager import HelperProcessManager
 from rlbot.base_extension import BaseExtension
-from rlbot.botmanager.agent_metadata import AgentMetadata
-from rlbot.botmanager.bot_helper_process import BotHelperProcess
 from rlbot.botmanager.bot_manager_flatbuffer import BotManagerFlatbuffer
 from rlbot.botmanager.bot_manager_independent import BotManagerIndependent
 from rlbot.botmanager.bot_manager_struct import BotManagerStruct
@@ -37,16 +36,16 @@ class SetupManager:
     parameters = None
     start_match_configuration = None
     agent_metadata_queue = None
-    agent_metadata_map = None
-    helper_process_map = None
-    quit_event = None
     extension = None
 
     def __init__(self):
         self.logger = get_logger(DEFAULT_LOGGER)
         self.game_interface = GameInterface(self.logger)
         self.quick_chat_manager = QuickChatManager(self.game_interface)
+        self.quit_event = mp.Event()
+        self.helper_process_manager = HelperProcessManager(self.quit_event)
         self.bot_quit_callbacks = []
+        self.agent_metadata_map = {}
 
     def startup(self):
         if self.has_started:
@@ -54,14 +53,7 @@ class SetupManager:
         self.logger.debug("Starting up game management")
         self.game_interface.inject_dll()
         self.game_interface.load_interface()
-        # Create Quit event
-        self.quit_event = mp.Event()
-
-        self.agent_metadata_map = {}
         self.agent_metadata_queue = mp.Queue()
-
-        self.helper_process_map = {}
-
         self.has_started = True
 
     def load_config(self, framework_config=None, config_location=DEFAULT_RLBOT_CONFIG_LOCATION, bot_configs=None, looks_configs=None):
@@ -125,7 +117,7 @@ class SetupManager:
                 break
             try:
                 single_agent_metadata = self.agent_metadata_queue.get(timeout=1)
-                self.start_or_update_helper_process(single_agent_metadata)
+                self.helper_process_manager.start_or_update_helper_process(single_agent_metadata)
                 self.agent_metadata_map[single_agent_metadata.index] = single_agent_metadata
                 configure_processes(self.agent_metadata_map, self.logger)
             except queue.Empty:
@@ -133,29 +125,6 @@ class SetupManager:
             except Exception as ex:
                 self.logger.error(ex)
                 pass
-
-    def start_or_update_helper_process(self, agent_metadata: AgentMetadata):
-        """
-        Examines the agent metadata to see if the agent needs a helper process. If the process is not running yet,
-        create the process. Once the process is running, feed the agent metadata to it.
-
-        If a process is created here, the pid will be added to the agent metadata.
-        """
-
-        helper_req = agent_metadata.helper_process_request
-
-        if helper_req is not None:
-            if helper_req.key not in self.helper_process_map:
-                metadata_queue = mp.Queue()
-                process = mp.Process(target=SetupManager.run_helper_process,
-                                     args=(helper_req.python_file_path, metadata_queue, self.quit_event))
-                process.start()
-                agent_metadata.pids.add(process.pid)
-
-                self.helper_process_map[helper_req.key] = metadata_queue
-
-            metadata_queue = self.helper_process_map[helper_req.key]
-            metadata_queue.put(agent_metadata)
 
     def shut_down(self):
         self.logger.info("Shutting Down")
@@ -192,16 +161,3 @@ class SetupManager:
             bm = BotManagerStruct(terminate_event, callback_event, config_file, name, team,
                                   index, agent_class_wrapper, agent_telemetry_queue, queue_holder)
         bm.run()
-
-    @staticmethod
-    def run_helper_process(python_file, metadata_queue, quit_event):
-        """
-        :param python_file: The absolute path of a python file containing the helper process that should be run.
-        It must define a class which is a subclass of BotHelperProcess.
-        :param metadata_queue: A queue from which the helper process will read AgentMetadata updates.
-        :param quit_event: An event which should be set when rlbot is shutting down.
-        """
-        class_wrapper = import_class_with_base(python_file, BotHelperProcess)
-        helper_class = class_wrapper.get_loaded_class()
-        helper = helper_class(metadata_queue, quit_event)
-        helper.start()
